requirements.txt
adaptivecardbuilder==0.0.9
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
coloredlogs==15.0.1
DateTime==5.2
Flask==2.3.2
frozenlist==1.3.3
future==0.18.3
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
schedule==1.2.0
tqdm==4.65.0
urllib3==2.0.3
webex-bot==0.4.0
webexteamssdk==1.6.1
websockets==10.2
Werkzeug==2.3.6
yarl==1.9.2
zope.interface==6.0
